requirements-dev.txt
# Development dependencies for local testing
flask==3.0.0
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
moto==4.2.11  # For mocking AWS services in tests
black==23.12.0
flake8==6.1.0